requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
flask-cors==4.0.0
gunicorn==21.2.0 